moving things into connection
