+++ appengine/clang_catalog.py
@@ -21,7 +21,7 @@ class ClangCatalog(Catalog):
             raise RuntimeError(
                 "no schema for relation %s because no connection" % rel_key)
         try:
-            dataset_info = self.check_datasets(relation_args)
+            dataset_info = self.connection.check_datasets(relation_args)
         except myria.MyriaError:
             raise ValueError('No relation {} in the catalog'.format(rel_key))
 
@@ -33,14 +33,6 @@ class ClangCatalog(Catalog):
         schema = {'columnNames': col_names, 'columnTypes': col_types}
 
         return scheme.Scheme(zip(schema['columnNames'], schema['columnTypes']))
-
-    def check_datasets(self, rel_args):
-        url = 'http://%s/catalog' % (self.connection.get_conn_string())
-        r = requests.Session().post(url, data=json.dumps(rel_args))
-        ret = r.json()
-        if ret:
-            return ret
-        raise myria.MyriaError
 
     def get_num_servers(self):
         if not self.connection:
@@ -58,14 +50,9 @@ class ClangCatalog(Catalog):
             raise RuntimeError(
                 "no cardinality of %s because no connection" % rel_key)
         try:
-            dataset_info = self.get_num_tuples(relation_args)
+            dataset_info = self.connection.get_num_tuples(relation_args)
         except myria.MyriaError:
             raise ValueError(rel_key)
         num_tuples = dataset_info['numTuples']
         assert type(num_tuples) is int
         return num_tuples
-
-    def get_num_tuples(self, rel_args):
-        url = 'http://%s/tuples' % (self.connection.get_conn_string())
-        r = requests.Session().post(url, data=json.dumps(rel_args))
-        return r.json()
